Extracting transform logic Adding schema to transformed file
