+++ src/Microsoft.VisualStudio.SlowCheetah.VS/Resources/Resources.resx
@@ -139,6 +139,7 @@
   </data>
   <data name="JsonTransformContents" xml:space="preserve">
     <value>{
+    "$schema": "http://json.schemastore.org/jdt"
 }
 </value>
     <comment>Contents of a newly created JSON transform file</comment>
