+++ src/Microsoft.VisualStudio.SlowCheetah.VS/Transform/TransformationPreviewLogger.cs
