+++ src/Microsoft.VisualStudio.SlowCheetah/Transformer/TransformerFactory.cs
@@ -5,6 +5,7 @@ namespace Microsoft.VisualStudio.SlowCheetah
 {
     using System;
     using System.IO;
+    using System.Xml;
 
     /// <summary>
     /// Factory for <see cref="ITransformer"/>
@@ -17,17 +18,57 @@ namespace Microsoft.VisualStudio.SlowCheetah
         /// <param name="source">Path to the file to be transformed</param>
         /// <param name="logger">Logger to be used in the transformer</param>
         /// <param name="useSections">Wheter or not to use sections while logging</param>
-        /// <returns>The instance of </returns>
+        /// <returns>The appropriate transformer for the given file</returns>
         public static ITransformer GetTransformer(string source, ITransformationLogger logger, bool useSections)
         {
             if (IsJsonFile(source))
             {
                 return new JsonTransformer(logger);
             }
-            else
+            else if (IsXmlFile(source))
             {
                 return new XmlTransformer(logger, useSections);
             }
+            else
+            {
+                throw new NotSupportedException($"{source} is not a supported file type for transformation");
+            }
+        }
+
+        /// <summary>
+        /// Verifies if a file is in XML format.
+        /// Attempts to open a file using an XML Reader.
+        /// </summary>
+        /// <param name="filepath">Full path to the file</param>
+        /// <returns>True is the file is XML</returns>
+        public static bool IsXmlFile(string filepath)
+        {
+            if (string.IsNullOrWhiteSpace(filepath))
+            {
+                throw new ArgumentNullException(nameof(filepath));
+            }
+
+            if (!File.Exists(filepath))
+            {
+                throw new FileNotFoundException("File not found", filepath);
+            }
+
+            bool isXmlFile = true;
+            try
+            {
+                using (XmlTextReader xmlTextReader = new XmlTextReader(filepath))
+                {
+                    // This is required because if the XML file has a DTD then it will try and download the DTD!
+                    xmlTextReader.DtdProcessing = DtdProcessing.Ignore;
+                    xmlTextReader.Read();
+                }
+            }
+            catch (XmlException)
+            {
+                isXmlFile = false;
+            }
+
+            return isXmlFile;
         }
 
         /// <summary>
@@ -38,6 +79,17 @@ namespace Microsoft.VisualStudio.SlowCheetah
         public static bool IsJsonFile(string filePath)
         {
             return Path.GetExtension(filePath).Equals(".json", StringComparison.OrdinalIgnoreCase);
+        }
+
+        /// <summary>
+        /// Verifies if a file is of a supported format.
+        /// JSON or XML
+        /// </summary>
+        /// <param name="filepath">Full path to the file</param>
+        /// <returns>True is the file type is supported</returns>
+        public static bool IsSupportedFile(string filepath)
+        {
+            return IsJsonFile(filepath) || IsXmlFile(filepath);
         }
     }
 }
